refact(ProductSelector): added foundation for redesign due travel/accident standalone
