+++ .env.example
@@ -4,6 +4,7 @@
 FEATURE_QUOTE_CART_API=SE,DK,NO
 FEATURE_OFFER_PAGE_INSURANCE_TOGGLE=SE
 FEATURE_OFFER_PAGE_PRODUCT_SELECTOR=NO
+FEATURE_OFFER_PAGE_TRAVEL_ACCIDENT_STANDALONE=DK,NO
 FEATURE_CHECKOUT_CREDIT_CHECK=NO
 FEATURE_CHECKOUT_UPSELL_CARD=SE
 FEATURE_CUSTOMER_SERVICE_PHONE_NUMBER=SE
