+++ src/client/pages/Offer/ProductSelector/NewProductSelector.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { QuoteCartQuery } from 'data/graphql'
+
+export type NewProductSelectorProps = { quoteCartQueryData: QuoteCartQuery }
+
+export const NewProductSelector = () => {
+  return <div>New Product Selector</div>
+}
