+++ src/shared/clientConfig.ts
@@ -8,6 +8,7 @@ export enum Feature {
   TEST_FEATURE = 'TEST_FEATURE', // For unit testing purposes
   OFFER_PAGE_INSURANCE_TOGGLE = 'OFFER_PAGE_INSURANCE_TOGGLE',
   OFFER_PAGE_PRODUCT_SELECTOR = 'OFFER_PAGE_PRODUCT_SELECTOR',
+  OFFER_PAGE_TRAVEL_ACCIDENT_STANDALONE = 'OFFER_PAGE_TRAVEL_ACCIDENT_STANDALONE',
   CHECKOUT_CREDIT_CHECK = 'CHECKOUT_CREDIT_CHECK',
   QUOTE_CART_API = 'QUOTE_CART_API',
   CHECKOUT_UPSELL_CARD = 'CHECKOUT_UPSELL_CARD',
